odometer_reading_detection/requirements.txt
easyocr==1.6.2
matplotlib==3.7.1
numpy==1.23.5
onnx==1.12.0
opencv-python-headless
pandas==1.4.4
Pillow==9.4.0
PyYAML==5.4.1
scipy==1.10.0
setuptools==65.6.3
torch==2.0.0
torchvision==0.15.1
wandb==0.15.0
scikit-image
python-bidi
PyYAML
Shapely
pyclipper
ninja
tqdm
seaborn
